% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient.R
\name{use_patient}
\alias{use_patient}
\title{Create a loading bar}
\usage{
use_patient(theme_name = NULL, theme_color = NULL, options = list())
}
\arguments{
\item{theme_name}{Name of the theme, see the list of choices below.}

\item{theme_color}{Name of the color, see the list of choices below. It is not possible to specify a color if \code{theme_name} is \code{NULL}. If \code{theme_name} is specified but \code{theme_color} is \code{NULL}, the default color is black.}
}
\description{
Create a loading bar
}
\details{
List of themes:
\itemize{
\item barber-shop
\item big-counter
\item bounce
\item center-atom
\item center-circle
\item center-radar
\item center-simple
\item corner-indicator
\item fill-left
\item flash
\item flat-top
\item loading-bar
\item mac-osx
\item material
\item minimal
}

List of colors:
\itemize{
\item black
\item blue
\item green
\item orange
\item pink
\item purple
\item red
\item silver
\item white
\item yellow
}
}
